Creation of project.
Decided on C# tailored comments that allow for automated
documentation with Doxygen, throughout the life of this project.

First tests for methods  that allow working with a mocked repository
and saving a mocked entity to that repository: successful.
Project fully complies with no errors or warnings.
All current tests pass.